Big commit mainly fixing some caching issues, where the esn prediction does not update when preprocessing steps are added.
Now predict and train are also returning the esn_obj again. This is necessary for proper updating.
Added a max_entries for all caches of 10. Otherwise the web app will run out of memory.
Removed the "clear cache" functionality, since it does not work for st.cache (which is unfortunatley needed for the esn_obj caching).
Added some more seed functionality for the time series preproceessing.
Added temp_seed to utils (might be the wrong place and not a good functionality but it works) for now.